improved motor class
Added new method that can read out the current speed. This is usefull when
initializing the MotorControl from a shared resource.

Fixed a bug, where setting the speed would only change the output voltage,
but would not log speed in class.

Changed speed from float to int. It's int anyway and float is a realy
wastefull variable on the arduino uno (no FPU + necessary type
conversions)