renamed files to make code compile on unix
includes are context sensitive on a Linux platform, but not on a
windows platform. It might be a different version of the ArduinoIDE
though. Fixed all those differences.

Changed folder and name of .ino from main to android_main, because main
is already taken in Linux version of the AndroidIDE.